[fix] (breaking) Search scope could end up in wrong arguments index
Previously, it was added at the end of the arguments, but that worked
only as long as the first three mandatory arguments were provided and
the optional ones were left at defaults.

This fixes this behaviour and, along the way, removes the search scope
argument completely when using the read(), list() and search()
methods. If full compliance with PHP's library is required,
ldapSearch() has been added - it has arguments signature identical to
native implementation.